in cumrate implementation, use int instead of size_t consistently

+++ include/omc/cumrate.h
@@ -113,7 +113,7 @@ public:
     /**
      * The value corresponding to each cumulated probability
      */
-    std::array<size_t,N> value;
+    std::array<int,N> value;
 
     /**
      * The cumulated probabilities (partial sums)
